requirements.txt
numpy
opencv-python
pillow
scikit-image
deskew
easyocr
torch
torchvision
spacy
nltk
